refactor(meal): 급식관련 이슈 리팩토링
