+++ src/templates/meal/constants/index.ts
@@ -1,3 +1,1 @@
-export { default as DATE } from "./date.constant";
-export { default as DIRECTION } from "./direction.constant";
 export { default as MEAL } from "./meal.constant";
